+++ src/nonebot_plugin_parser/parsers/bilibili/__init__.py
@@ -252,12 +252,12 @@ class BilibiliParser(BaseParser):
         # 下载封面
         if cover := room_data.cover:
             cover_task = self.downloader.download_img(cover, ext_headers=self.headers)
-            contents.append(ImageContent(cover_task))
+            contents.append(self.create_image_content(cover_task))
 
         # 下载关键帧
         if keyframe := room_data.keyframe:
             keyframe_task = self.downloader.download_img(keyframe, ext_headers=self.headers)
-            contents.append(ImageContent(keyframe_task))
+            contents.append(self.create_image_content(keyframe_task))
 
         author = self.create_author(room_data.name, room_data.avatar)
